december commit

+++ controllers/sites.js
@@ -13,9 +13,7 @@ export const createSite = async (req, res) => {
         [name, siteType, address, latitude, longitude]
       );
 
-      const [rows2] = await pool.query(
-        `UPDATE sites SET siteId = 1000+id`
-      );
+      const [rows2] = await pool.query(`UPDATE sites SET siteId = 1000+id`);
 
       const [rows3] = await pool.query(
         `INSERT INTO crudLogs (userId, action, tableName, columnName, entryId, oldData, newData) VALUES (?, ?, ?, ?, ?, ?, ?)`,
@@ -55,28 +53,34 @@ export const getSite = async (req, res) => {
 
 export const downloadSitePdf = async (req, res) => {
   try {
-    const doc = new PDFDocument({ margin: 30, size: 'A4' });
-    const filename = 'sites_list.pdf';
+    const doc = new PDFDocument({ margin: 30, size: "A4" });
+    const filename = "sites_list.pdf";
     const pdfStream = doc.pipe(fs.createWriteStream(filename));
 
     // PDF metadata
-    doc.info['Title'] = 'Site Details';
+    doc.info["Title"] = "Site Details";
 
     // Fetch data from the database
-    const [rows] = await pool.query('SELECT * FROM sites');
+    const [rows] = await pool.query("SELECT * FROM sites");
 
     // Create table data
-    let count=1;
+    let count = 1;
     const table = {
-      title: 'Site Details',
+      title: "Site Details",
       headers: [
-        { label:"Sl.No", property: 'slNo', width: 50, renderer: null },
-        { label:"Site Id", property: 'id', width: 50, renderer: null },
-        { label:"Site Name", property: 'name', width: 150, renderer: null }, 
-        { label:"Site Type", property: 'type', width: 100, renderer: null }, 
-        { label:"Address", property: 'address', width: 200, renderer: null },
+        { label: "Sl.No", property: "slNo", width: 50, renderer: null },
+        { label: "Site Id", property: "id", width: 50, renderer: null },
+        { label: "Site Name", property: "name", width: 150, renderer: null },
+        { label: "Site Type", property: "type", width: 100, renderer: null },
+        { label: "Address", property: "address", width: 200, renderer: null },
       ],
-      rows: rows.map((row) => [count++, row.id, row.name, row.siteType, row.address]),
+      rows: rows.map((row) => [
+        count++,
+        row.id,
+        row.name,
+        row.siteType,
+        row.address,
+      ]),
     };
 
     // Calculate the horizontal center position for the table
@@ -91,11 +95,11 @@ export const downloadSitePdf = async (req, res) => {
     });
 
     // Respond to the request with the PDF as a download
-    res.setHeader('Content-Type', 'application/pdf');
-    res.setHeader('Content-Disposition', `attachment; filename="${filename}"`);
+    res.setHeader("Content-Type", "application/pdf");
+    res.setHeader("Content-Disposition", `attachment; filename="${filename}"`);
 
     // Pipe the PDF stream to the response
-    pdfStream.on('finish', () => {
+    pdfStream.on("finish", () => {
       fs.createReadStream(filename).pipe(res);
     });
 
@@ -103,7 +107,7 @@ export const downloadSitePdf = async (req, res) => {
     doc.end();
   } catch (err) {
     console.error(err);
-    res.status(500).send('Something broke!');
+    res.status(500).send("Something broke!");
   }
 };
 
@@ -114,10 +118,9 @@ export const updateSite = async (req, res) => {
     console.log(req.body);
 
     if (name && siteType) {
-      const [rows0] = await pool.query(
-        `SELECT * FROM sites WHERE id = ?`,
-        [req.params.id]
-      );
+      const [rows0] = await pool.query(`SELECT * FROM sites WHERE id = ?`, [
+        req.params.id,
+      ]);
 
       const [rows, fields] = await pool.query(
         `UPDATE sites SET name = ?, siteType = ?, address = ?, latitude = ?, longitude = ? WHERE id = ?`,
@@ -137,17 +140,33 @@ export const updateSite = async (req, res) => {
       if (prevSiteType !== siteType) {
         const [rows1] = await pool.query(
           `INSERT INTO crudLogs (userId, action, tableName, columnName, entryId, oldData, newData) VALUES (?, ?, ?, ?, ?, ?, ?)`,
-          [userId, "update", "sites", "siteType", req.params.id, prevSiteType, siteType]
+          [
+            userId,
+            "update",
+            "sites",
+            "siteType",
+            req.params.id,
+            prevSiteType,
+            siteType,
+          ]
         );
       }
       if (prevAddress !== address) {
         const [rows1] = await pool.query(
           `INSERT INTO crudLogs (userId, action, tableName, columnName, entryId, oldData, newData) VALUES (?, ?, ?, ?, ?, ?, ?)`,
-          [userId, "update", "sites", "address", req.params.id, prevAddress, address]
+          [
+            userId,
+            "update",
+            "sites",
+            "address",
+            req.params.id,
+            prevAddress,
+            address,
+          ]
         );
       }
 
-      res.status(201).json({ id: req.params.id});
+      res.status(201).json({ id: req.params.id });
     }
   } catch (err) {
     console.log(err);
@@ -159,6 +178,29 @@ export const updateSite = async (req, res) => {
 export const deleteSite = async (req, res) => {
   try {
     const { userId, name } = req.body;
+
+    const [users] = await pool.query(
+      `SELECT id, name, userRole FROM users WHERE siteId = ?`,
+      [req.params.id]
+    );
+    if (users.length > 0) {
+      res
+        .status(403)
+        .send(
+          `site has currently assigned with ${users[0].userRole} ${users[0].name} (userId: ${users[0].id})`
+        );
+      return;
+    }
+
+    const [stocks] = await pool.query(
+      `SELECT id FROM stocks WHERE siteId = ?`,
+      [req.params.id]
+    );
+    if (stocks.length > 0) {
+      res.status(403).send(`site still has some stocks left`);
+      return;
+    }
+
     const [rows, fields] = await pool.query(`DELETE FROM sites WHERE id = ?`, [
       req.params.id,
     ]);
@@ -170,7 +212,7 @@ export const deleteSite = async (req, res) => {
       [userId, "delete", "sites", "name", rows.insertId, name, ""]
     );
 
-    res.status(200).json({status: "success"});
+    res.status(200).json({ status: "success" });
   } catch (err) {
     console.log(err);
     res.status(500).send("Something broke!");
